requirements.txt
dataclasses
google-cloud-bigquery
python-dotenv